refactor(acp): isolate session configuration
